+++ crates/core/CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Clone` for `Container` ([#666](https://github.com/stac-utils/stac-rs/pull/666))
 - `Serialize` for `Container` ([#667](https://github.com/stac-utils/stac-rs/pull/667))
 
+### Changed
+
+- Default to snappy compression for geoparquet ([#673](https://github.com/stac-utils/stac-rs/pull/673))
+
 ## [0.12.0] - 2025-01-31
 
 ### Added